# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PayReality Runtime Authority API",
    "description": "Deterministic policy evaluation (OPA/Rego), ED25519-signed Evidence, and the human-review resolution flow for AI agent financial actions. Full schema at /openapi.json, interactive docs at /docs.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/intents": {
      "post": {
        "tags": [
          "intents"
        ],
        "summary": "Submit Intent",
        "description": "spec 19.5. The `agent` dependency has already verified the request\nsignature over the raw body; this handler is authenticated by the\ntime it runs.",
        "operationId": "submit_intent_v1_intents_post",
        "parameters": [
          {
            "name": "x-payreality-key-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Payreality-Key-Id"
            }
          },
          {
            "name": "x-payreality-signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Payreality-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitIntentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitIntentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/decisions/{decision_id}": {
      "get": {
        "tags": [
          "intents"
        ],
        "summary": "Get Decision",
        "description": "New (not in spec 19's literal API): the poll endpoint a caller uses\nuntil a HUMAN_REVIEW decision is resolved (see plan's addition).",
        "operationId": "get_decision_v1_decisions__decision_id__get",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Decision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/decisions/{decision_id}/resolve": {
      "post": {
        "tags": [
          "intents"
        ],
        "summary": "Resolve Decision",
        "description": "The Phase 1 addition (see plan's 'The one addition: resolving\nHUMAN_REVIEW'). Gated by the shared operator key (app.security) so not\nanyone can resolve a review; resolved_by is still a free-text identity\nstring until the real human RBAC system in the V3 roadmap replaces it.",
        "operationId": "resolve_decision_v1_decisions__decision_id__resolve_post",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Decision Id"
            }
          },
          {
            "name": "x-payreality-operator-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Payreality-Operator-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/evidence/{evidence_id}": {
      "get": {
        "tags": [
          "evidence"
        ],
        "summary": "Get Evidence",
        "description": "spec 19.6.",
        "operationId": "get_evidence_v1_evidence__evidence_id__get",
        "parameters": [
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/evidence": {
      "get": {
        "tags": [
          "evidence"
        ],
        "summary": "List Evidence",
        "description": "spec 19.6.",
        "operationId": "list_evidence_v1_evidence_get",
        "parameters": [
          {
            "name": "decision_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvidenceResponse"
                  },
                  "title": "Response List Evidence V1 Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/evidence/{evidence_id}/verify": {
      "post": {
        "tags": [
          "evidence"
        ],
        "summary": "Verify Evidence",
        "description": "spec 19.7 / 17.5. A False result indicates tampering or corruption\nand must be treated as a P1 operational incident by the caller.",
        "operationId": "verify_evidence_v1_evidence__evidence_id__verify_post",
        "parameters": [
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEvidenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DecisionSummary": {
        "properties": {
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "decision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Decision Id"
          },
          "evaluated_mandates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evaluated Mandates"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "outcome",
          "decision_id",
          "evaluated_mandates"
        ],
        "title": "DecisionSummary"
      },
...
      "EvidenceResponse": {
        "properties": {
          "evidence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidence Id"
          },
          "decision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Decision Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "key_id": {
            "type": "string",
            "title": "Key Id"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "evidence_id",
          "decision_id",
          "payload",
          "key_id",
          "signature",
          "status",
          "created_at"
        ],
        "title": "EvidenceResponse",
        "description": "spec 17.1 + 19.6."
      },
      "GetDecisionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "evaluated_mandates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evaluated Mandates"
          },
          "resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "outcome",
          "reason",
          "agent_id",
          "action",
          "amount",
          "currency",
          "evaluated_mandates"
        ],
        "title": "GetDecisionResponse",
        "description": "New (not in spec 19's literal API): the polling endpoint the\nHUMAN_REVIEW-resolution addition needs (see plan)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolutionSummary": {
        "properties": {
          "resolution": {
            "type": "string",
            "title": "Resolution"
          },
          "resolved_by": {
            "type": "string",
            "title": "Resolved By"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "resolution",
          "resolved_by",
          "created_at"
        ],
        "title": "ResolutionSummary"
      },
      "ResolveDecisionRequest": {
        "properties": {
          "resolution": {
            "type": "string",
            "title": "Resolution"
          },
          "resolved_by": {
            "type": "string",
            "title": "Resolved By"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "resolution",
          "resolved_by"
        ],
        "title": "ResolveDecisionRequest",
        "description": "The Phase 1 addition (see plan) that resolves a HUMAN_REVIEW decision\nwithout mutating the immutable Decision row."
      },
      "ResolveDecisionResponse": {
        "properties": {
          "decision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Decision Id"
          },
          "resolution": {
            "$ref": "#/components/schemas/ResolutionSummary"
          },
          "evidence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidence Id"
          }
        },
        "type": "object",
        "required": [
          "decision_id",
          "resolution",
          "evidence_id"
        ],
        "title": "ResolveDecisionResponse"
      },
...
      "SubmitIntentRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "counterparty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "nonce": {
            "type": "string",
            "title": "Nonce"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "action",
          "amount",
          "currency",
          "requested_at",
          "nonce"
        ],
        "title": "SubmitIntentRequest",
        "description": "spec 19.5, extended with `nonce` (required by spec 9.3/21.2's replay\nprotection, present in the field table but omitted from the section\n19.5 example) and `correlation_id` (spec 9.3, optional)."
      },
      "SubmitIntentResponse": {
        "properties": {
          "intent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Intent Id"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionSummary"
          },
          "evidence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidence Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "intent_id",
          "decision",
          "evidence_id",
          "status"
        ],
        "title": "SubmitIntentResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerifyEvidenceResponse": {
        "properties": {
          "evidence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidence Id"
          },
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "verified_at": {
            "type": "string",
            "format": "date-time",
            "title": "Verified At"
          },
          "key_id": {
            "type": "string",
            "title": "Key Id"
          }
        },
        "type": "object",
        "required": [
          "evidence_id",
          "valid",
          "verified_at",
          "key_id"
        ],
        "title": "VerifyEvidenceResponse",
        "description": "spec 19.7."
      }
    }
  }
}